Рефакторинг: переместить асинхронные методы в соответствующий класс
